Moved IndexNow into a service folder as indexnow-ping
no ref

Relocates the flat single-file service to services/indexnow-ping,
matching the explore-ping naming for outbound ping services, and renames
listen() to init() to match the boot lifecycle convention used by the
other services. Pure move: no behaviour or logic changes.
